Some small fixes to make Minion compile on Windows again
git-svn-id: svn+ssh://keith.cs.st-and.ac.uk/mnt/raid/svn/repositories/minion-private/public/trunk@846 477b24dd-e91e-0410-aa49-980f3df0729b

+++ minion/constraints/constraint_pow.h
@@ -67,13 +67,13 @@ struct PowConstraint : public Constraint
     
   
   double my_pow(DomainInt x, DomainInt y)
-  { return pow(checked_cast<int>(x), checked_cast<int>(y));}
+  { return pow((double)checked_cast<int>(x), checked_cast<int>(y));}
   
   double my_y(DomainInt x, DomainInt z)
-  { return log(checked_cast<int>(z)) / log(checked_cast<int>(x)); }
+  { return log((double)checked_cast<int>(z)) / log((double)checked_cast<int>(x)); }
   
   double my_x(DomainInt y, DomainInt z)
-  { return exp(log(checked_cast<int>(z)) / checked_cast<int>(y)); }
+  { return exp(log((double)checked_cast<int>(z)) / checked_cast<int>(y)); }
   
   PROPAGATE_FUNCTION(int flag, DomainDelta)
   {
